refact: renomear pacotes e ajustar referências de classes para unificação do namespace
