+++ sus_integrated/src/main/java/br/com/susintegrated/model/Address.java
@@ -1,0 +1,63 @@
+package br.com.susintegrated.model;
+
+import br.com.susintegrated.controller.dto.address.AddressFormDTO;
+
+public class Address {
+
+    private String street;
+    private Integer number;
+    private String neighborhood;
+    private String city;
+    private String state;
+    private String zipCode;
+
+    public Address(String street, Integer number, String neighborhood, String city, String state, String zipCode) {
+        this.street = street;
+        this.number = number;
+        this.neighborhood = neighborhood;
+        this.city = city;
+        this.state = state;
+        this.zipCode = zipCode;
+    }
+
+    public Address(AddressFormDTO formDTO) {
+        this.neighborhood = formDTO.neighborhood();
+        this.city = formDTO.city();
+        this.state = formDTO.state();
+        this.number = formDTO.number();
+        this.street = formDTO.street();
+        this.zipCode = formDTO.zipCode();
+    }
+
+    public void merge(AddressFormDTO formDTO) {
+        this.neighborhood = formDTO.neighborhood();
+        this.city = formDTO.city();
+        this.state = formDTO.state();
+        this.number = formDTO.number();
+        this.street = formDTO.street();
+    }
+
+    public String getStreet() {
+        return street;
+    }
+
+    public Integer getNumber() {
+        return number;
+    }
+
+    public String getNeighborhood() {
+        return neighborhood;
+    }
+
+    public String getCity() {
+        return city;
+    }
+
+    public String getState() {
+        return state;
+    }
+
+    public String getZipCode() {
+        return zipCode;
+    }
+}
