+++ sus_integrated/src/main/java/br/com/susintegrated/repository/PatientRepository.java
@@ -1,10 +1,9 @@
 package br.com.susintegrated.repository;
 
-import br.com.susintegrated.model.patient.Patient;
+import br.com.susintegrated.model.Patient;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
