+++ sus_scheduling/src/main/java/br/com/sus_scheduling/SusScheduling.java
@@ -1,0 +1,17 @@
+package br.com.sus_scheduling;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableAsync;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+@SpringBootApplication
+@EnableScheduling
+@EnableAsync
+public class SusScheduling {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SusScheduling.class, args);
+	}
+
+}
