+++ sus_scheduling/src/main/java/br/com/sus_scheduling/config/SwaggerConfig.java
@@ -1,0 +1,20 @@
+package br.com.sus_scheduling.config;
+
+
+import org.springdoc.core.models.GroupedOpenApi;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class SwaggerConfig {
+
+    @Bean
+    GroupedOpenApi publicApi(){
+        return GroupedOpenApi
+                .builder()
+                .group("public-apis")
+                .pathsToMatch("/**")
+                .build();
+    }
+
+}
