+++ sus_scheduling/src/main/java/br/com/sus_scheduling/exception/PatientException.java
@@ -1,0 +1,7 @@
+package br.com.sus_scheduling.exception;
+
+public class PatientException extends RuntimeException {
+    public PatientException(String message) {
+        super(message);
+    }
+}
