+++ sus_scheduling/src/main/java/br/com/sus_scheduling/exception/RescheduleException.java
@@ -1,0 +1,9 @@
+package br.com.sus_scheduling.exception;
+
+public class RescheduleException extends RuntimeException {
+
+    public RescheduleException(String message) {
+        super(message);
+    }
+
+}
