+++ sus_scheduling/src/main/java/br/com/sus_scheduling/model/SchedulingStatus.java
@@ -1,0 +1,22 @@
+package br.com.sus_scheduling.model;
+
+public enum SchedulingStatus {
+
+    UNDER_ANALYSIS("Em análise"),
+    SCHEDULED("Agendado"),
+    RESCHEDULED("Reagendado"),
+    CANCELED("Cancelado"),
+    DONE("Realizado"),
+    ATTENDING("Atendendo"),
+    LATE("Atrasado");
+
+    private final String description;
+
+    SchedulingStatus(String description) {
+        this.description = description;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+}
