+++ sus_scheduling/src/main/java/br/com/sus_scheduling/queue/consumer/MessageBodyByIntegrated.java
@@ -1,0 +1,7 @@
+package br.com.sus_scheduling.queue.consumer;
+
+import java.util.UUID;
+
+public record MessageBodyByIntegrated(boolean validatedPatient, UUID schedulingId) {
+
+}
