+++ sus_scheduling/src/main/java/br/com/sus_scheduling/queue/producer/MessageBodyForIntegrated.java
@@ -1,0 +1,10 @@
+package br.com.sus_scheduling.queue.producer;
+
+import java.io.Serializable;
+import java.util.UUID;
+
+public record MessageBodyForIntegrated(UUID patientId, UUID schedulingId) implements Serializable {
+
+    private static final long serialVersionUID = 1L;
+
+}
