+++ sus_scheduling/src/main/java/br/com/sus_scheduling/repository/SchedulingRepository.java
@@ -1,0 +1,34 @@
+package br.com.sus_scheduling.repository;
+
+import br.com.sus_scheduling.model.Scheduling;
+import br.com.sus_scheduling.model.SchedulingStatus;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+import org.springframework.stereotype.Repository;
+
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.UUID;
+
+@Repository
+public interface SchedulingRepository extends JpaRepository<Scheduling, UUID> {
+
+    List<Scheduling> findAllByPatientId(UUID patientId);
+
+    List<Scheduling> findAllByProfessionalId(UUID id);
+
+    @Query(value = "SELECT * FROM Scheduling s WHERE CAST(s.appointment AS DATE) = :date", nativeQuery = true)
+    List<Scheduling> findAllByAppointmentsToDay(@Param("date") LocalDate date);
+
+    List<Scheduling> findAllByStatus(SchedulingStatus status);
+
+    List<Scheduling> findAllByAppointmentBetweenAndStatusIn(LocalDateTime now, LocalDateTime next24Hours, List<SchedulingStatus> statusList);
+
+    @Query(value = "SELECT * FROM Scheduling s WHERE s.patient_id =:id and CAST(s.appointment AS DATE) = :date", nativeQuery = true)
+    List<Scheduling> findAllByPatientIdAndDate(UUID id, LocalDate date);
+
+    @Query(value = "SELECT * FROM Scheduling s WHERE s.professional_id =:id and CAST(s.appointment AS DATE) = :date", nativeQuery = true)
+    List<Scheduling> findAllByProfessionalIdAndDate(UUID id, LocalDate date);
+}
