--- sus_scheduling/src/main/java/br/com/susscheduling/SusScheduling.java
@@ -1,17 +1,0 @@
-package br.com.susscheduling;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.scheduling.annotation.EnableScheduling;
-
-@SpringBootApplication
-@EnableScheduling
-@EnableAsync
-public class SusScheduling {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SusScheduling.class, args);
-	}
-
-}
