--- sus_scheduling/src/main/java/br/com/susscheduling/exception/PatientException.java
@@ -1,7 +1,0 @@
-package br.com.susscheduling.exception;
-
-public class PatientException extends RuntimeException {
-    public PatientException(String message) {
-        super(message);
-    }
-}
