--- sus_scheduling/src/main/java/br/com/susscheduling/exception/RescheduleException.java
@@ -1,9 +1,0 @@
-package br.com.susscheduling.exception;
-
-public class RescheduleException extends RuntimeException {
-
-    public RescheduleException(String message) {
-        super(message);
-    }
-
-}
