--- sus_scheduling/src/main/java/br/com/susscheduling/model/scheduling/SchedulingStatus.java
@@ -1,22 +1,0 @@
-package br.com.susscheduling.model.scheduling;
-
-public enum SchedulingStatus {
-
-    UNDER_ANALYSIS("Em análise"),
-    SCHEDULED("Agendado"),
-    RESCHEDULED("Reagendado"),
-    CANCELED("Cancelado"),
-    DONE("Realizado"),
-    ATTENDING("Atendendo"),
-    LATE("Atrasado");
-
-    private final String description;
-
-    SchedulingStatus(String description) {
-        this.description = description;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-}
