--- sus_scheduling/src/main/java/br/com/susscheduling/queue/consumer/MessageBodyByIntegrated.java
@@ -1,7 +1,0 @@
-package br.com.susscheduling.queue.consumer;
-
-import java.util.UUID;
-
-public record MessageBodyByIntegrated(boolean validatedPatient, UUID schedulingId) {
-
-}
