--- sus_scheduling/src/main/java/br/com/susscheduling/queue/consumer/MessageBodyByUnity.java
@@ -1,13 +1,0 @@
-package br.com.susscheduling.queue.consumer;
-
-import java.io.Serializable;
-import java.util.UUID;
-
-public record MessageBodyByUnity(boolean validatedSpeciality, boolean validatedUnity, UUID schedulingId) implements Serializable {
-
-    private static final long serialVersionUID = 1L;
-
-    public boolean isValidated() {
-        return validatedSpeciality && validatedUnity;
-    }
-}
