--- sus_scheduling/src/main/java/br/com/susscheduling/queue/producer/MessageBodyForIntegrated.java
@@ -1,10 +1,0 @@
-package br.com.susscheduling.queue.producer;
-
-import java.io.Serializable;
-import java.util.UUID;
-
-public record MessageBodyForIntegrated(UUID patientId, UUID schedulingId) implements Serializable {
-
-    private static final long serialVersionUID = 1L;
-
-}
