--- sus_scheduling/src/main/java/br/com/susscheduling/queue/producer/MessageProducer.java
@@ -1,34 +1,0 @@
-package br.com.susscheduling.queue.producer;
-
-import br.com.susscheduling.config.RabbitConfig;
-import org.springframework.amqp.core.Message;
-import org.springframework.amqp.core.MessageProperties;
-import org.springframework.amqp.rabbit.core.RabbitTemplate;
-import org.springframework.amqp.support.converter.Jackson2JsonMessageConverter;
-import org.springframework.stereotype.Service;
-
-@Service
-public class MessageProducer {
-
-    private final RabbitTemplate rabbitTemplate;
-    private final Jackson2JsonMessageConverter jackson2JsonMessageConverter;
-
-    public MessageProducer(RabbitTemplate rabbitTemplate, Jackson2JsonMessageConverter jackson2JsonMessageConverter) {
-        this.rabbitTemplate = rabbitTemplate;
-        this.jackson2JsonMessageConverter = jackson2JsonMessageConverter;
-    }
-
-    public void sendToUnity(MessageBodyForUnity message) {
-        MessageProperties messageProperties = new MessageProperties();
-        messageProperties.setContentType(MessageProperties.CONTENT_TYPE_JSON);
-        Message rabbitMessage = jackson2JsonMessageConverter.toMessage(message, messageProperties);
-        rabbitTemplate.convertAndSend(RabbitConfig.EXCHANGE_NAME, RabbitConfig.ROUTING_KEY_SCHEDULING_UNITY, rabbitMessage);
-    }
-
-    public void sendToIntegrated(MessageBodyForIntegrated message) {
-        MessageProperties messageProperties = new MessageProperties();
-        messageProperties.setContentType(MessageProperties.CONTENT_TYPE_JSON);
-        Message rabbitMessage = jackson2JsonMessageConverter.toMessage(message, messageProperties);
-        rabbitTemplate.convertAndSend(RabbitConfig.EXCHANGE_NAME, RabbitConfig.ROUTING_KEY_SCHEDULING_INTEGRATED, rabbitMessage);
-    }
-}
